Merge nuts.davemloft.net:/disk1/BK/network-2.6 into nuts.davemloft.net:/disk1/BK/net-2.6
2004/07/10 12:13:58-07:00 yoshfuji
[NET]: Fix dst_underflow_bug_msg printk args.

2004/07/10 12:13:01-07:00 herbert
[IPCOMP6]: Exclude IPCOMP header from props.header_len

Now that the IPv4 encap stuff is out of the way, I'll be sending you
the IPv6 versions.

Here is the one to remove the unnecessary extra space reserved for
IPCOMP.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Yoshifuji Hideaki <yoshfuji@linux-ipv6.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/10 11:53:57-07:00 herbert
[IPSEC]: Move generic encap code into xfrm4_output

I've finally finished merging the general encapsulation code for IPv4.
Here is the patch.

The idea is basically to make x->type->output similar in structure to
x->type->input.  That means moving the tunnel encapsulation and other
generic code out.

They have ended up in xfrm4_output.c.

The advantage of this is that we have exactly one copy of the tunnel
encapsulation code.  So if we need to change it (e.g., set the TTL
according to the route) then it's easier and less error-prone.

In fact, in doing so I've already noticed that the ECN wasn't being
copied correctly in everything except xfrm4_tunnel.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/09 17:06:40-07:00 davem
Merge http://linux-mh.bkbits.net/bluetooth-2.6
into nuts.davemloft.net:/disk1/BK/net-2.6

2004/07/09 17:05:22-07:00 tomd
[CRYPTO]: Set CRYPTO_TFM_RES_BAD_KEY_LEN in twofish.

2004/07/09 16:56:37-07:00 herbert
[IPSEC]: Fix uh->len when doing NATT with IP options.

I just noticed that the UDP header length in esp4_output() is incorrect
when IP options are present (in transport mode).  This patch fixes exactly
that.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/09 16:51:35-07:00 herbert
[XFRM]: Add FLUSHSA and FLUSHPOLICY.

This patch adds FLUSHSA and FLUSHPOLICY to xfrm_user which are
analagous to SADB_FLUSH and SADB_X_SPDFLUSH in af_key.

This is useful in KMs on startup/shutdown so that the system is
reset to a known state.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/09 16:50:29-07:00 jmorris
[CRYPTO]: Remove lazy allocation from deflate.

Signed-off-by: James Morris <jmorris@redhat.com>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/09 16:46:16-07:00 shemminger
[PKT_SCHED]: Add jitter support to netem.

This patch adds jitter if desired to the delayed packets in the
netem scheduler.  I dropped the rate stuff out and reorganized so
that an underlying pfifo queue is used (next plan is to make it
have class ops).

The jitter is given as sigma to a Gaussian normal distribution.  The actual
implementation is a reduced form of the table driven stuff in NISTnet
(free).

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/09 16:37:16-07:00 shemminger
[BRIDGE]: Support different MTU sizes.

This patch adds support for different size MTU's to bridging.
It is useful for bridging Ethernet's with jumbo frames, etc.

The mtu of the bridge pseudo-device is maintained as the minimum
of all the underlying ports.  And when forwarding a frame through
the bridge, it will drop the frame if the outgoing port's MTU
is less than the frame size (as per 802 standard).

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/09 16:26:07-07:00 davem
[PKT_SCHED]: Remove CSZ scheduler.

It was an experimental hack and never finished off.

Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/09 23:15:32+02:00 marcel
[Bluetooth] Don't reset the USB halted bits

The USB halted bits are useless and so there is no need to reset them
after changing the configuration.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/07/09 23:12:28+02:00 marcel
[Bluetooth] Respond to L2CAP info requests

This patch adds the correct behavior when a L2CAP info request is sent
by the peer. The answer should be a L2CAP info response with the result
code set to "not supported".

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/07/08 16:34:29-07:00 davem
[IPV4]: Fix multicast socket hangs.

If a multicast packet gets looped back, the sending
socket can hang if a local read just sits and does
not empty its receive queue.

The problem is that when an SKB clone is freed up,
the destructor is only invoked for the head SKB when
there is a fraglist (which is created for fragmentation).

The solution is to account the fragment list SKB lengths
in the top-level head SKB, then it all works out.

Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/07 15:22:10-07:00 jheffner
[TCP]: Do not round window to MSS if window scaling.

Signed-off-by: John Heffner <jheffner@psc.edu>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/07 15:20:08-07:00 davem
[TCP]: Type qualifiers, such as const, are ignored on function return type.

2004/07/07 15:19:10-07:00 hadi
[PKT_SCHED]: Another missed tc_stats spinlock conversion.

Signed-off-by: Jamal Hadi Salim <hadi@znyx.com>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/07 14:06:11-07:00 hirofumi
[NET]: Cleanup mis-usage of seq_release_private

These doesn't allocate memory and doesn't use seq->private. However
kfree() ignores NULL, so these are not the problem.

This patch just cleans these up.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/07 14:05:03-07:00 hirofumi
[IPV4]: IPMR fixes.

- pim_protocol warning fix

- ipmr_vif_open() and ipmr_mfc_open() allocates the memory, so it
  should use seq_release_private().

- ipmr_mfc_seq_xxx is using it->cache, in order to control whether
  unlock should be do or not, but it->cache was not initialized in
  ipmr_mfc_seq_start(). So it can point the previous state if user
  did seek(). This become to the cause of twice unlock.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/07 14:02:02-07:00 shemminger
[NET]: Deinline sock_i_uid, sock_i_ino

The sock_i_uid and sock_i_ino functions are only called by
/proc type interfaces, so they don't need to be inlined.

Also, the inline functions writeable, rcvtimeo, sndtimeo are test
for value functions that don't change their argument.

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/07 14:00:14-07:00 kaber
[NETFILTER]: Fix two broken checks for options in ipt_LOG

Signed-off-by: Patrick McHardy <kaber@trash.net>
Signed-off-by: Harald Welte <laforge@netfilter.org>
Signed-off-by: David S. Miller <davem@redhat.com>

BKrev: 40f1d4c5EsJyg07Xwb8ccDK6CWLI5A