[PATCH] compat_fillonedir() warning fix
	access_ok() expects a pointer, not unsigned long.  It's not a
problem on platforms that have this guy done as a macro (or ones that
do not use fs/compat.c at all), but that's still wrong and on some
platforms that care we actually have access_ok() as inlined function.
Bogus cast removed.

BKrev: 40f42a39RjNHh8kqx57MPuTLNUA9SQ