[PATCH] Fix memory corruption at x86-64 SMP bootup
This fixes a long standing corruption bug in the x86-64 code. The SMP
trampoline would corrupt page 2, which was sometimes used for other
data.  This corrupted the ioport/iomem list in some cases and causes
oopses while read /proc/iomem.

Fix is to check the correct location and don't zero it afterwards
because it gets reinitialized for the next CPU anyways.

Thanks to Alexander Nyberg for tracking it down.

BKrev: 40f9f322NstojzLb3OK2SxvOIbe0yg