Merge http://linux-mh.bkbits.net/bluetooth-2.6 into nuts.davemloft.net:/disk1/BK/net-2.6
2004/07/16 14:05:36+02:00 marcel
[Bluetooth] Use a signed integer for the RSSI value

The RSSI value in the inquiry response with RSSI must be a signed
integer and not an unsigned one.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/07/16 13:27:29+02:00 marcel
[Bluetooth] Add missing entry for the HIDP support

This small patch adds the missing entry about the HIDP support
to the main Kconfig file of the Bluetooth subsystem.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/07/16 12:59:11+02:00 marcel
[Bluetooth] Make use of usb_kill_urb()

Now that usb_kill_urb() is in the main kernel tree it should
be used. This patch makes the needed modifications to the USB
Bluetooth driver.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/07/16 12:53:17+02:00 marcel
[Bluetooth] Add support for another ALPS module

This patch adds the specific vendor and product id's for another
ALPS module which don't uses the USB Bluetooth class id.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

BKrev: 40feeb22J7xNRECxQSE3cs3jsyC8uw