[INET]: Create enum of ECN bits
This patch is a preparation for an update of the ECN encap/decap
code with respect to RFC3168.

It creates an enum of the four code-points defined by RFC3168
and uses them throughout the inet_ecn.h file.

The only non-trivial bit is in IP_ECN_set_ce/IP6_ECN_set_ce where
the patch uses INET_ECN_CE instead of 1.  This is OK as those
functions assume that the ECT bit is already set.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>

Signed-off-by: David S. Miller <davem@redhat.com>

BKrev: 41001283DKeUz1VeipEVYBphdM9ACA