Merge davem@nuts.davemloft.net:/disk1/BK/net-2.6 into kernel.bkbits.net:/home/davem/net-2.6
2004/07/24 22:41:08-07:00 davem
[SCTP]: Fix mis-merge.

2004/07/24 22:31:11-07:00 anton
[NET]: Use NET_IP_ALIGN in acenic.

Use NET_IP_ALIGN in acenic driver. Also remove the 16 byte padding,
caches can be anywhere from 16 to 256 bytes and the skb should be
cacheline aligned already.

Signed-off-by: Anton Blanchard <anton@samba.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/24 22:26:27-07:00 herbert
[AH6]: Replace skb by iph in clear_mutable_options.

This patch replaces the skb argument in ipv6_clear_mutable_options() by
an ipv6hdr.  Doing so allows us to point skb->nh elsewhere when calling
this function.

I've also thrown in some obvious clean-ups for that function.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/24 22:24:48-07:00 herbert
[AH4]: Save daddr iff options are present.

This is a little optimisation for AH4.  When I moved the tunnel code out,
I put the daddr copying code on the main path which is unnecessary since
daddr is only mutable if IP options are present.

This patch moves the saving and restoring of daddr under the check for
the existence of IP options.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/24 22:19:43-07:00 davem
Cset exclude: shemminger@osdl.org|ChangeSet|20040722205059|21273

2004/07/24 22:17:10-07:00 davem
Merge conflicts with Yoshfuji's SNMP stats changes.

2004/07/24 22:01:14-07:00 kaber
[IPV4/IPV6]: Add myself to MAINTAINERS.

2004/07/24 16:54:25-07:00 davem
Merge nuts.davemloft.net:/disk1/BK/network-2.6
into nuts.davemloft.net:/disk1/BK/net-2.6

2004/07/23 16:41:31-07:00 sri
[SCTP] Mark chunks as ineligible for fast retransmit after they are
retransmitted. Also mark any chunks that could not be fit in the
PMTU sized packet as ineligible for fast retransmit.

Signed-off-by: Sridhar Samudrala <sri@us.ibm.com>

2004/07/23 14:14:25-07:00 shemminger
[NET]: eql sparse cleanup.

Minor 0 vs NULL cleanup

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/23 14:12:36-07:00 shemminger
[IPVS]: Convert to module_param.

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/23 13:46:55-07:00 davem
[PKT_SCHED]: Alpha not studly enough for SCH_CLK_CPU.

The cycle counter it provides overflows very quickly,
on the order of 10 minutes, so it is not suitable for
this purpose.

Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/23 13:45:21-07:00 kaber
[PKT_SCHED]: Make clock source configurable

Signed-off-by: Patrick McHardy <kaber@trash.net>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/23 13:26:52-07:00 herbert
[AH6]: Disallow mutable bits after AH header.

As we discussed before, mutable headers should not be allowed after
the AH header.  In fact, this appears to be the intention of RFC 2402.
It is further clarified in section 3.1.1 of

http://www.ietf.org/internet-drafts/draft-ietf-ipsec-rfc2402bis-07.txt

This allows us to simplify the code in ah6.c.  As a result, this also
fixes the following issues:

* Dependence on skb->h in ah6_output().
* Bogus clearing of auth_data of 2nd AH header in ipv6_clear_mutable_options().

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/23 13:23:48-07:00 herbert
[IPSEC]: Missing unlock in policy timer.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/23 13:08:38-07:00 davem
Merge nuts.davemloft.net:/disk1/BK/network-2.6
into nuts.davemloft.net:/disk1/BK/net-2.6

2004/07/22 23:22:01-07:00 sri
[SCTP] Fix missing '+' in the computation of sack chunk size in
sctp_sm_pull_sack().

Signed-off-by: Jorge Hernandez <jhh@lucent.com>
Signed-off-by: Sridhar Samudrala <sri@us.ibm.com>

2004/07/22 23:18:24-07:00 sri
[SCTP] Use idr_get_new_above() with a starting id of 1 to avoid returning
an associd of 0.

Signed-off-by: Sridhar Samudrala <sri@us.ibm.com>

2004/07/22 23:15:55-07:00 sri
[SCTP] Fix issues with handling stale cookie error over multihoming
associations.

Signed-off-by: Jorge Hernandez <jhh@lucent.com>
Signed-off-by: Sridhar Samudrala <sri@us.ibm.com>

2004/07/22 23:13:05-07:00 sri
[SCTP] Fix data not being delivered to user in SHUTDOWN_SENT state.

Also cleaned up sctp_sf_eat_data_6_2() and sctp_sf_eat_data_fast_4_4()
as they have a lot of common code.

Signed-off-by: Jorge Hernandez <jhh@lucent.com>
Signed-off-by: Sridhar Samudrala <sri@us.ibm.com>

2004/07/22 23:09:04-07:00 sri
[SCTP] Set/Get default SCTP_PEER_ADDR_PARAMS for endpoint when associd
and peer address are 0.

Signed-off-by: Anand R. Setlur <asetlur@lucent.com>
Signed-off-by: Sridhar Samudrala <sri@us.ibm.com>

2004/07/22 17:20:03-07:00 kaber
[PKT_SCHED]: Use get_cycles() for PSCHED_CPU clock source

Signed-off-by: Patrick McHardy <kaber@trash.net>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/22 17:13:11-07:00 kaber
[PKT_SCHED]: Remove dead timer code.

Signed-off-by: Patrick McHardy <kaber@trash.net>
Signed-off-by: David S. Miller <davem@redhat.com>

BKrev: 4103b55e7TnJY8uA24qV0gbF2xvWgQ