[PATCH] BIO page refcounting fix
Hopefully fixes the free-of-a-freed-page BUG caused during CDRW writing.

This also fixes a problem in the bouncing for io errors (it needs to free
the pages and clear the BIO_UPTODATE flag, not set it.  it's already set.
passing -EIO to bio_endio() takes care of that).

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 410891d2WRA8DoLQKOMFrhQ9ZnPTbQ