[PATCH] ppc64: cpu hotplug fix
On Power4 and earlier hardware there is no need to clear the CPPR (see RPAp
479 section 18.5.4.7.2 for what little info there is on the CPPR) when
stopping a cpu.  On hardware that uses Power5 an undocumented change has
been made that requires the CPPR to be cleared if an isolate is to be done
on the stopped cpu.  So the following patch lets cpu hotplug work on the
recent hardware.

I sent this patch to the ppc64-dev list back in mid April and Suse picked
it up then for SLES9 so it has been well tested for several months.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4108922cV_oMiMjnun4czhR_egOpEg