[PATCH] ppc32: fix compilation with binutils-2.15
Currently, ppc32 will not always compile with binutils-2.15.  The issue
is that binutils has become even more strict about which opcodes can be
used with which CPU flags.  The problem is that we have a number of
cases where we compile with altivec instructions (with runtime checks to
make sure we can actually run them) in code that's not altivec specific.

The fix for this is to always pass in -maltivec on CONFIG_6xx.  To do
this cleanly, we split our AFLAGS definition up into
aflags-$(CONFIG_FOO).

Signed-off-by: Tom Rini <trini@kernel.crashing.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 410c09derQ7j5o10-kQuNuAqh5ZzCA