[PATCH] quieten down per-zone memory stats
On a system with a lot of nodes, 4 lines of output per node is a lot to
have to sit through as the system comes up, especially if you're on the
other end of a slow serial link.  The information is valuable though, so
keep it around for the system logger.  This patch makes the printks for the
memory stats use KERN_DEBUG instead of the default loglevel.

Signed-off-by: Jesse Barnes <jbarnes@sgi.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 410e75f5ktHEx-R7Gz2MRgHStxL8jA