[PATCH] fix /proc printing of TASK_DEAD state
I just stumbled across this patch that's been sitting in my tree for ages.

I thought I'd sent this in before.  It's a trivial fix for the printing
of task state in /proc and sysrq dumps and such, so that TASK_DEAD shows
up correctly.  This state is pretty much only ever there to be seen when
there are exit/reaping bugs, but it's not like that hasn't come up.

Signed-off-by: Roland McGrath <roland@redhat.com>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4112d447fJclz_IoNw1NK2edUu4Tjw