[PATCH] [ppc64] Fix SLB castout issue
The SLB rewrite removed a fix for a hard to hit bug, but the SFS guys
managed to hit it straight away. We need to check both r1 and PACAKSAVE
or else we could cast our kernel segment out when on the irq or softirq
stack.

Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4117e51baUc9yAlv-irP3QtmQXa61Q