Merge http://linux-watchdog.bkbits.net/linux-2.6-watchdog into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/08/20 11:18:39+02:00 wim
[WATCHDOG] v2.6.8.1 watchdog-llseek-patch

The watchdog drivers use a VFS implementation and thus should not be
lseek'able, so we put a '.llseek = no_llseek' in the file_operations
structure.

2004/08/20 10:55:20+02:00 wim
[WATCHDOG] v2.6.8.1 cpu5wdt.c-nonseekable_open-patch

cpu5wdt also contains a VFS and thus should be "nonseekable_open"

2004/08/20 09:02:36+02:00 fl
[WATCHDOG] pcwd-watchdog.txt-patch

Fix example program in pcwd-watchdog.txt document.

2004/08/20 08:29:39+02:00 arnd
[WATCHDOG] v2.6.8.1 compat_ioctl-patch

The watchdog ioctl interface is defined correctly for 32 bit emulation,
although WIOC_GETSUPPORT was not marked as such, for an unclear reason.
WDIOC_SETTIMEOUT and WDIOC_GETTIMEOUT were added in may 2002 to the
code but never to the ioctl list. This adds all three definitions.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Wim Van Sebroeck <wim@iguana.be>

BKrev: 412952acAuvvKvY-6RnT_U8X2HN-Fg